+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyTemporary/2-Update.ps1
@@ -1,0 +1,43 @@
+<#
+This example is used to test new resources and showcase the usage of new resources being worked on.
+It is not meant to use as a production baseline.
+#>
+
+Configuration Example
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [PSCredential]
+        $credsCredential
+    )
+    Import-DscResource -ModuleName Microsoft365DSC
+
+    Node localhost
+    {
+        AADAuthenticationMethodPolicyTemporary "AADAuthenticationMethodPolicyTemporary-TemporaryAccessPass"
+        {
+            Credential               = $Credscredential;
+            DefaultLength            = 9; # Updated Property
+            DefaultLifetimeInMinutes = 60;
+            Ensure                   = "Present";
+            ExcludeTargets           = @(
+                MSFT_AADAuthenticationMethodPolicyTemporaryExcludeTarget{
+                    Id = 'All Company'
+                    TargetType = 'group'
+                }
+            );
+            Id                       = "TemporaryAccessPass";
+            IncludeTargets           = @(
+                MSFT_AADAuthenticationMethodPolicyTemporaryIncludeTarget{
+                    Id = 'DSCGroup'
+                    TargetType = 'group'
+                }
+            );
+            IsUsableOnce             = $False;
+            MaximumLifetimeInMinutes = 480;
+            MinimumLifetimeInMinutes = 60;
+            State                    = "enabled";
+        }
+    }
+}