+++ Modules/Microsoft365DSC/Examples/Resources/IntuneDeviceConfigurationPolicyAndroidOpenSourceProject/2-Update.ps1
@@ -1,0 +1,35 @@
+<#
+This example is used to test new resources and showcase the usage of new resources being worked on.
+It is not meant to use as a production baseline.
+#>
+
+Configuration Example
+{
+    param(
+        [Parameter(Mandatory = $true)]
+        [PSCredential]
+        $Credscredential
+    )
+    Import-DscResource -ModuleName Microsoft365DSC
+
+    node localhost
+    {
+        IntuneDeviceConfigurationPolicyAndroidOpenSourceProject 'myAndroidOpenSourceProjectPolicy'
+        {
+            DisplayName               = 'aosp'
+            Assignments               = @(
+                MSFT_DeviceManagementConfigurationPolicyAssignments {
+                    deviceAndAppManagementAssignmentFilterType = 'none'
+                    dataType                                   = '#microsoft.graph.allDevicesAssignmentTarget'
+                }
+            )
+            CameraBlocked             = $True # Updated Property
+            FactoryResetBlocked       = $True
+            PasswordRequiredType      = 'deviceDefault'
+            ScreenCaptureBlocked      = $True
+            StorageBlockExternalMedia = $True
+            Ensure                    = 'Present'
+            Credential                = $Credscredential
+        }
+    }
+}