+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -1019,8 +1019,7 @@ function New-IntuneSettingCatalogPolicy
 
     try
     {
-	$BaseUrl = $Global:MSCloudLoginConnectionProfile.Intune.GraphBaseUrl
-        $Uri = '$($BaseUrl)/beta/deviceManagement/configurationPolicies'
+        $Uri = '/beta/deviceManagement/configurationPolicies'
 
         $policy = @{
             'name'              = $Name
@@ -1083,8 +1082,7 @@ function Update-IntuneSettingCatalogPolicy
 
     try
     {
-        $BaseUrl = $Global:MSCloudLoginConnectionProfile.Intune.GraphBaseUrl
-        $Uri = "$($BaseUrl)/beta/deviceManagement/configurationPolicies/$DeviceConfigurationPolicyId"
+        $Uri = "/beta/deviceManagement/configurationPolicies/$DeviceConfigurationPolicyId"
 
         $policy = @{
             'name'              = $Name
@@ -1284,11 +1282,11 @@ function Update-DeviceConfigurationPolicyAssignment
         [System.String]
         $APIVersion = 'beta'
     )
+
     try
     {
         $deviceManagementPolicyAssignments = @()
-        $BaseUrl = $Global:MSCloudLoginConnectionProfile.Intune.GraphBaseUrl
-        $Uri = "$($BaseUrl)/$APIVersion/$Repository/$DeviceConfigurationPolicyId/assign"
+        $Uri = "/$APIVersion/$Repository/$DeviceConfigurationPolicyId/assign"
 
         foreach ($target in $targets)
         {
@@ -1323,6 +1321,69 @@ function Update-DeviceConfigurationPolicyAssignment
             -TenantId $TenantId `
             -Credential $Credential
 
+        return $null
+    }
+}
+
+function Get-OmaSettingPlainTextValue
+{
+    [CmdletBinding()]
+    [OutputType([System.String])]
+    Param(
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $SecretReferenceValueId,
+
+        [Parameter()]
+        [ValidateSet('v1.0', 'beta')]
+        [System.String]
+        $APIVersion = 'beta'
+    )
+
+    try
+    {
+        <#
+            e.g. PolicyId for SecretReferenceValueId '35ea58ec-2a79-471d-8eea-7e28e6cd2722_bdf6c690-05fb-4d02-835d-5a7406c35d58_abe32712-2255-445f-a35e-0c6f143d82ca'
+            is 'bdf6c690-05fb-4d02-835d-5a7406c35d58'
+        #>
+        $SplitSecretReferenceValueId = $SecretReferenceValueId.Split("_")
+        if ($SplitSecretReferenceValueId.Count -eq 3)
+        {
+            $PolicyId = $SplitSecretReferenceValueId[1]
+        }
+        else
+        {
+            return $null
+        }
+    }
+    catch
+    {
+        return $null
+    }
+
+    $Repository = 'deviceManagement/deviceConfigurations'
+    $Uri = "/{0}/{1}/{2}/getOmaSettingPlainTextValue(secretReferenceValueId='{3}')" -f $APIVersion, $Repository, $PolicyId, $SecretReferenceValueId
+
+    try
+    {
+        $Result = Invoke-MgGraphRequest -Method GET -Uri $Uri -ErrorAction Stop
+    }
+    catch
+    {
+        $Message = "Error decrypting OmaSetting with SecretReferenceValueId {0}:" -f $SecretReferenceValueId
+        New-M365DSCLogEntry -Message $Message `
+            -Exception $_ `
+            -Source $($MyInvocation.MyCommand.Source) `
+            -TenantId $TenantId `
+            -Credential $Credential
+
+        return $null
+    }
+
+    if (![String]::IsNullOrEmpty($Result.Value))
+    {
+        return $Result.Value
+    } else {
         return $null
     }
 }